Move admin to admin.rs
